Refactor function calls
