+++ src/sema/statements.rs
@@ -5,13 +5,17 @@ use super::contracts::is_base;
 use super::diagnostics::Diagnostics;
 use super::eval::check_term_for_constant_overflow;
 use super::expression::{
-    available_functions, call_expr, expression, function_call_expr, function_call_pos_args,
-    named_call_expr, named_function_call_expr, ExprContext, ResolveTo,
+    expression,
+    function_call::{available_functions, call_expr, named_call_expr},
+    ExprContext, ResolveTo,
 };
 use super::symtable::{LoopScopes, Symtable};
 use crate::sema::builtin;
 use crate::sema::expression::constructor::{
     constructor_named_args, match_constructor_to_args, new,
+};
+use crate::sema::expression::function_call::{
+    function_call_expr, function_call_pos_args, named_function_call_expr,
 };
 use crate::sema::function_annotation::function_body_annotations;
 use crate::sema::symtable::{VariableInitializer, VariableUsage};
